Add a bar

+++ app/models/model.py
@@ -114,5 +114,5 @@ class GPT2PPL:
             prev_end_loc = end_loc
             if end_loc == seq_len:
                 break
-        ppl = int(torch.exp(torch.stack(nlls).sum() / end_loc))
+        ppl = float(torch.exp(torch.stack(nlls).sum() / end_loc))
         return ppl
